Fast cut of separate cache cleanup
