+++ flowmachine/setup.py
@@ -58,7 +58,10 @@ setup(
     version=versioneer.get_version(),
     cmdclass=versioneer.get_cmdclass(),
     entry_points={
-        "console_scripts": ["flowmachine = flowmachine.core.server.server:main"]
+        "console_scripts": [
+            "flowmachine = flowmachine.core.server.server:main",
+            "cache-cleanup = flowmachine.core.server.cache_cleanup:main",
+        ]
     },
     description="Digestion program for Call Detail Record (CDR) data.",
     long_description=readme,
